Add description to poll questions translations
So we can show it as a hint for the question when defined.

We needed to recall `globalize_accessors` from the
custom `Poll::Question` model. Otherwise the application fails
when calling the globalize accessor for the new `description`
translatable field. The weirdest thing is that only fails in the
test `spec/system/admin/translatable_spec.rb:389`. It can be
reproduced locally and when running CI. The application
seems to work fine in the development environment