Use a different action to export polls results and allow administrators to do it
The admin poll show action should be always accesible no matter the status of
the poll. However the export feature should only be availabe once the poll is
done, just like the poll results.

As the request format is not available in the abilities context we would need to
change the abilities implementation to pass in the format. In place of doing that,
I'm creating a new action to export the poll results within de admin poll results
controller which has more sense to me, so now we can fully control permissions
without needing the request format.